# docs/SERVER_MODE/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "MrRSS is a modern, cross-platform desktop RSS reader with auto-translation, smart feed discovery, and AI-powered summarization.",
        "title": "MrRSS API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/WCY-dt/MrRSS",
            "email": "mail@ch3nyang.top"
        },
        "license": {
            "name": "GPL-3.0",
            "url": "https://www.gnu.org/licenses/gpl-3.0.en.html"
        },
        "version": "1.3.19"
    },
    "host": "localhost:1234",
    "basePath": "/api",
    "paths": {
...
        "/feeds": {
            "get": {
                "description": "Retrieve all RSS feed subscriptions (passwords are cleared)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Get all feeds",
                "responses": {
                    "200": {
                        "description": "List of feeds",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Feed"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/feeds/add": {
            "post": {
                "description": "Add a new RSS/Atom/Email/Script/XPath feed subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Add a new feed",
                "parameters": [
                    {
                        "description": "Feed details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed added successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Feed URL already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/feeds/delete": {
            "post": {
                "description": "Delete a feed subscription by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Delete a feed",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/feeds/refresh": {
            "post": {
                "description": "Trigger a refresh for a specific feed (runs in background with progress tracking)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Refresh a single feed",
                "parameters": [
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Feed ID",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed refresh started successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request (invalid feed ID)",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Feed not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/feeds/reorder": {
            "post": {
                "description": "Change the position and optionally the category of a feed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Reorder a feed",
                "parameters": [
                    {
                        "description": "Reorder details (feed_id, category, position)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reorder status",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/feeds/update": {
            "post": {
                "description": "Update properties of an existing feed subscription",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "feeds"
                ],
                "summary": "Update a feed",
                "parameters": [
                    {
                        "description": "Feed update details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Feed updated successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Feed URL already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Feed": {
            "type": "object",
            "properties": {
                "article_view_mode": {
                    "description": "Article view mode override ('global', 'webpage', 'rendered')",
                    "type": "string"
                },
                "articles_per_month": {
                    "description": "Average articles per month (last 90 days / 3)",
                    "type": "number"
                },
                "auto_expand_content": {
                    "description": "Auto expand content mode ('global', 'enabled', 'disabled')",
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "discovery_completed": {
                    "description": "Track if discovery has been run",
                    "type": "boolean"
                },
                "email_address": {
                    "description": "Email/Newsletter support",
                    "type": "string"
                },
                "email_folder": {
                    "description": "IMAP folder to monitor (default INBOX)",
                    "type": "string"
                },
                "email_imap_port": {
                    "description": "IMAP server port (default 993)",
                    "type": "integer"
                },
                "email_imap_server": {
                    "description": "IMAP server address",
                    "type": "string"
                },
                "email_last_uid": {
                    "description": "Last processed email UID for incremental updates",
                    "type": "integer"
                },
                "email_password": {
                    "description": "IMAP password (encrypted)",
                    "type": "string"
                },
                "email_username": {
                    "description": "IMAP username",
                    "type": "string"
                },
                "freshrss_stream_id": {
                    "description": "FreshRSS stream ID (e.g., \"feed/http://...\")",
                    "type": "string"
                },
                "hide_from_timeline": {
                    "description": "Hide articles from timeline views",
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "image_url": {
                    "description": "New field",
                    "type": "string"
                },
                "is_freshrss_source": {
                    "description": "FreshRSS integration",
                    "type": "boolean"
                },
                "is_image_mode": {
                    "description": "Whether this feed is for image gallery mode",
                    "type": "boolean"
                },
                "last_error": {
                    "description": "Track last fetch error",
                    "type": "string"
                },
                "last_update_status": {
                    "description": "Last update status (\"success\" or \"failed\")",
                    "type": "string"
                },
                "last_updated": {
                    "type": "string"
                },
                "latest_article_time": {
                    "description": "Statistics",
                    "type": "string"
                },
                "link": {
                    "description": "Website homepage link",
                    "type": "string"
                },
                "position": {
                    "description": "Position within category for custom ordering",
                    "type": "integer"
                },
                "proxy_enabled": {
                    "description": "Whether to use proxy for this feed",
                    "type": "boolean"
                },
                "proxy_url": {
                    "description": "Custom proxy URL for this feed (overrides global)",
                    "type": "string"
                },
                "refresh_interval": {
                    "description": "Custom refresh interval in minutes (0 = use global, -1 = intelligent, -2 = never, \u003e0 = custom minutes)",
                    "type": "integer"
                },
                "script_path": {
                    "description": "Path to custom script for fetching feed",
                    "type": "string"
                },
                "tags": {
                    "description": "Tags (populated by API handlers)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Tag"
                    }
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "description": "XPath support for HTML/XML scraping",
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "xpath_item": {
                    "description": "XPath to extract feed items",
                    "type": "string"
                },
                "xpath_item_author": {
                    "description": "XPath to extract item author",
                    "type": "string"
                },
                "xpath_item_categories": {
                    "description": "XPath to extract item categories",
                    "type": "string"
                },
                "xpath_item_content": {
                    "description": "XPath to extract item content",
                    "type": "string"
                },
                "xpath_item_thumbnail": {
                    "description": "XPath to extract item thumbnail",
                    "type": "string"
                },
                "xpath_item_time_format": {
                    "description": "Time format for parsing timestamp",
                    "type": "string"
                },
                "xpath_item_timestamp": {
                    "description": "XPath to extract item timestamp",
                    "type": "string"
                },
                "xpath_item_title": {
                    "description": "XPath to extract item title",
                    "type": "string"
                },
                "xpath_item_uid": {
                    "description": "XPath to extract item unique ID",
                    "type": "string"
                },
                "xpath_item_uri": {
                    "description": "XPath to extract item URI",
                    "type": "string"
                }
            }
        },
...
        "models.Tag": {
            "type": "object",
            "properties": {
                "color": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "position": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}